requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via ml-server (pyproject.toml)
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   jupyter-server
    #   starlette
argon2-cffi==23.1.0
    # via jupyter-server
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
arrow==1.3.0
    # via isoduration
asttokens==3.0.0
    # via stack-data
async-lru==2.0.5
    # via jupyterlab
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
babel==2.17.0
    # via jupyterlab-server
beautifulsoup4==4.13.4
    # via nbconvert
bleach==6.2.0
    # via nbconvert
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via argon2-cffi-bindings
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via uvicorn
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
exceptiongroup==1.2.2
    # via
    #   anyio
    #   ipython
executing==2.2.0
    # via stack-data
fastapi==0.115.12
    # via ml-server (pyproject.toml)
fastjsonschema==2.21.1
    # via nbformat
fqdn==1.5.1
    # via jsonschema
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via jupyterlab
idna==3.10
    # via
    #   anyio
    #   httpx
    #   jsonschema
    #   requests
    #   yarl
ipykernel==6.29.5
    # via
    #   jupyter
    #   jupyter-console
    #   jupyterlab
ipython==8.36.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==8.1.7
    # via jupyter
isoduration==20.11.0
    # via jsonschema
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
joblib==1.5.0
    # via scikit-learn
json5==0.12.0
    # via jupyterlab-server
jsonpointer==3.0.0
    # via jsonschema
jsonschema==4.23.0
    # via
    #   jupyter-events
    #   jupyterlab-server
    #   nbformat
jsonschema-specifications==2025.4.1
    # via jsonschema
jupyter==1.1.1
    # via ml-server (pyproject.toml)
jupyter-client==8.6.3
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
jupyter-console==6.6.3
    # via jupyter
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
    #   jupyterlab
    #   nbclient
    #   nbconvert
    #   nbformat
jupyter-events==0.12.0
    # via jupyter-server
jupyter-lsp==2.2.5
    # via jupyterlab
jupyter-server==2.15.0
    # via
    #   jupyter-lsp
    #   jupyterlab
    #   jupyterlab-server
    #   notebook
    #   notebook-shim
jupyter-server-terminals==0.5.3
    # via jupyter-server
jupyterlab==4.4.2
    # via
    #   jupyter
    #   notebook
jupyterlab-pygments==0.3.0
    # via nbconvert
jupyterlab-server==2.27.3
    # via
    #   jupyterlab
    #   notebook
jupyterlab-widgets==3.0.15
    # via ipywidgets
markupsafe==3.0.2
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mistune==3.1.3
    # via nbconvert
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
nbclient==0.10.2
    # via nbconvert
nbconvert==7.16.6
    # via
    #   jupyter
    #   jupyter-server
nbformat==5.10.4
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
nest-asyncio==1.6.0
    # via ipykernel
notebook==7.4.2
    # via jupyter
notebook-shim==0.2.4
    # via
    #   jupyterlab
    #   notebook
numpy==2.2.5
    # via
    #   scikit-learn
    #   scipy
overrides==7.7.0
    # via jupyter-server
packaging==25.0
    # via
    #   ipykernel
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
pandocfilters==1.5.1
    # via nbconvert
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
platformdirs==4.3.7
    # via jupyter-core
prometheus-client==0.21.1
    # via jupyter-server
prompt-toolkit==3.0.51
    # via
    #   ipython
    #   jupyter-console
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.3
    # via stack-data
pycparser==2.22
    # via cffi
pydantic==2.11.4
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via
    #   ipython
    #   jupyter-console
    #   nbconvert
python-dateutil==2.9.0.post0
    # via
    #   arrow
    #   jupyter-client
python-dotenv==1.1.0
    # via ml-server (pyproject.toml)
python-json-logger==3.3.0
    # via jupyter-events
pyyaml==6.0.2
    # via jupyter-events
pyzmq==26.4.0
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-console
    #   jupyter-server
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   jupyter-events
requests==2.32.3
    # via
    #   ml-server (pyproject.toml)
    #   jupyterlab-server
rfc3339-validator==0.1.4
    # via
    #   jsonschema
    #   jupyter-events
rfc3986-validator==0.1.1
    # via
    #   jsonschema
    #   jupyter-events
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.6.1
    # via ml-server (pyproject.toml)
scipy==1.15.2
    # via scikit-learn
send2trash==1.8.3
    # via jupyter-server
setuptools==80.3.1
    # via jupyterlab
six==1.17.0
    # via
    #   python-dateutil
    #   rfc3339-validator
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
stack-data==0.6.3
    # via ipython
starlette==0.46.2
    # via fastapi
terminado==0.18.1
    # via
    #   jupyter-server
    #   jupyter-server-terminals
threadpoolctl==3.6.0
    # via scikit-learn
tinycss2==1.4.0
    # via bleach
tomli==2.2.1
    # via jupyterlab
tornado==6.4.2
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-console
    #   jupyter-core
    #   jupyter-events
    #   jupyter-server
    #   jupyterlab
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
types-python-dateutil==2.9.0.20241206
    # via arrow
typing-extensions==4.13.2
    # via
    #   anyio
    #   async-lru
    #   beautifulsoup4
    #   fastapi
    #   ipython
    #   mistune
    #   multidict
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typing-inspection
    #   uvicorn
typing-inspection==0.4.0
    # via pydantic
uri-template==1.3.0
    # via jsonschema
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via ml-server (pyproject.toml)
wcwidth==0.2.13
    # via prompt-toolkit
webcolors==24.11.1
    # via jsonschema
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.8.0
    # via jupyter-server
widgetsnbextension==4.0.14
    # via ipywidgets
yarl==1.20.0
    # via aiohttp
